Removed some warnings, fixed fault tolerance code and install travis CI

+++ distributedcombigrid/src/sgpp/distributedcombigrid/mpi_fault_simulator/Sim_FT_comm_free.cpp
@@ -39,7 +39,7 @@ int simft::Sim_FT_MPI_Comm_free(simft::Sim_FT_MPI_Comm *f_comm) {
       (*f_comm)->NBC_Vector_Send = simft::NBC_to_Vector((*f_comm));
     }
   }
-  for (int i = 0; i < (*f_comm)->NBC_Vector_Send.size(); i++) {
+  for (int i = 0; i < static_cast<int>((*f_comm)->NBC_Vector_Send.size()); i++) {
     std::cout << "NBCOUT:::" << (*f_comm)->NBC_Vector_Send[i] << "\n";
   }
 
@@ -107,7 +107,7 @@ int simft::Sim_FT_MPI_Comm_free2(simft::Sim_FT_MPI_Comm *f_comm) {  // does not 
       (*f_comm)->NBC_Vector_Send = simft::NBC_to_Vector((*f_comm));
     }
   }
-  for (int i = 0; i < (*f_comm)->NBC_Vector_Send.size(); i++) {
+  for (int i = 0; i < static_cast<int>((*f_comm)->NBC_Vector_Send.size()); i++) {
     std::cout << "NBCOUT:::" << (*f_comm)->NBC_Vector_Send[i] << "\n";
   }
